Quando um voluntario for editado, agora verifica as novas datas, e compara com escalas anteriormente criadas baseadas nas datas disponiveis antigas, caso as novas nao estejam de acordo, esse voluntario será removido da escala em que estava

+++ src/main/java/com/crescer_aprender/escala/service/VoluntarioService.java
@@ -1,9 +1,12 @@
 package com.crescer_aprender.escala.service;
 
+import com.crescer_aprender.escala.entity.Escala;
 import com.crescer_aprender.escala.entity.Voluntario;
 import com.crescer_aprender.escala.exception.ConstantExceptionUtil;
 import com.crescer_aprender.escala.exception.EntityNotFoundException;
 import com.crescer_aprender.escala.exception.InvalidVoluntarioDataException;
+import com.crescer_aprender.escala.exception.VoluntarioIsScheduledException;
+import com.crescer_aprender.escala.repository.EscalaRepository;
 import com.crescer_aprender.escala.repository.VoluntarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,10 +19,12 @@ import java.util.Optional;
 public class VoluntarioService {
 
     private final VoluntarioRepository repository;
+    private final EscalaRepository escalaRepository;
 
     @Autowired
-    public VoluntarioService(VoluntarioRepository repository) {
+    public VoluntarioService(VoluntarioRepository repository, EscalaRepository escalaRepository) {
         this.repository = repository;
+        this.escalaRepository = escalaRepository;
     }
 
     public Voluntario save(Voluntario voluntario) {
@@ -45,6 +50,17 @@ public class VoluntarioService {
         Optional.ofNullable(voluntario.getSenha()).ifPresent(oldVoluntario::setSenha);
         Optional.ofNullable(voluntario.getDatasDisponiveis()).ifPresent(dates -> mergeDatasDisponiveis(oldVoluntario, dates));
 
+        Optional<List<Escala>> escalasDoVoluntario = escalaRepository.findByVoluntario(oldVoluntario);
+
+        if(escalasDoVoluntario.isPresent()){
+            for (Escala escala : escalasDoVoluntario.get()) {
+                // Remover voluntário da escala caso não esteja disponível para alguma das datas da escala
+                escala.getVoluntarios().removeIf(v -> v.getId().equals(oldVoluntario.getId()) &&
+                        escala.getDatas().stream().anyMatch(date -> !oldVoluntario.getDatasDisponiveis().contains(date)));
+                escalaRepository.save(escala);
+            }
+
+        }
         return repository.save(oldVoluntario);
     }
 
@@ -61,6 +77,9 @@ public class VoluntarioService {
 
     public boolean delete(Long id) {
         if (repository.existsById(id)) {
+            if(escalaRepository.existsByVoluntarios(Voluntario.builder().id(id).build())){
+                throw new VoluntarioIsScheduledException();
+            }
             repository.deleteById(id);
             return true;
         } else {
